step5按登出
step6把local storage的”姓名“、“密碼” 刪除
＊此時首頁右上角只顯示「會員註冊/登入」，不顯示「會員姓名,購物車,會員中心,登出」＊

+++ NCU_MIS_SA/WEB-INF/classes/ncu/im3069/demo/app/OrderHelper.java
@@ -26,7 +26,7 @@ public class OrderHelper {
     public JSONObject create(Order order) {
         /** 記錄實際執行之SQL指令 */
         String exexcute_sql = "";
-        long idtbl_order = -1;
+        int idtbl_order = -1;
         JSONArray opa = new JSONArray();
         
         try {
@@ -69,7 +69,7 @@ public class OrderHelper {
             ResultSet rs = pres.getGeneratedKeys();
 
             if (rs.next()) {
-                idtbl_order = rs.getLong(1);
+                idtbl_order = rs.getInt(1);
                 ArrayList<OrderProduct> opd = order.getOrderProduct();
                 opa = oph.createByList(idtbl_order, opd);
             }
